refactor: 배럴 파일 수정
