+++ src/app/(main)/mypage/components/layout/ProfileSideBar.tsx
@@ -1,7 +1,8 @@
 'use client';
+
 import { Button, ProfileImg, ReviewItemBar } from '@/components/Atoms';
+import { useModal } from '@/components/Molecules';
 import { useAuthStore } from '@/apis/auth/hooks/authStore';
-import { useModal } from '@/components/Molecules/modal';
 import { ProfileEditModal } from './profileModal/ProfileEditModal';
 import { REVIEW_KEYWORD_LABELS } from '@/constants';
 import { useReviewStats } from '@/hooks';
