+++ src/app/(main)/shopping/components/UpdateShoppingForm.tsx
@@ -1,6 +1,14 @@
 'use client';
 
-import { Button, Dropdown, Label, Textarea, TextInput } from '@/components';
+import {
+  Button,
+  Label,
+  Textarea,
+  TextInput,
+  useToast,
+  KeywordChip,
+} from '@/components/Atoms';
+import { Dropdown } from '@/components/Molecules';
 import {
   CAPACITY_OPTIONS,
   MODEL_PROVINCE_OPTIONS,
@@ -15,7 +23,6 @@ import { useForm } from 'react-hook-form';
 import z from 'zod';
 import { useMutation } from '@tanstack/react-query';
 import { ApiResponse } from '@/types/common';
-import { useToast, KeywordChip } from '@/components/Atoms';
 import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 import { MeetingDetailType } from '@/types/meetingsType';
