+++ src/components/Atoms/index.ts
@@ -3,33 +3,35 @@ export {
   type ActionMenuProps,
   type ActionMenuItem,
 } from './ActionMenu/ActionMenu';
-export { Badge } from './Badge';
-export { Button } from './Button';
-export { Chip } from './Chip';
-export { DateFilter, type DateFilterProps } from './DateFilter';
+export { Badge } from './Badge/Badge';
+export { Button } from './Button/Button';
+export { Chip } from './Chip/Chip';
+export { DateFilter, type DateFilterProps } from './DateFilter/DateFilter';
 export { Icon } from './Icon/Icon';
 export type { IconType, IconSize } from './Icon/iconType';
-export { Label } from './Label';
+export { Label } from './Label/Label';
 export { LinkRenderer } from './LinkRenderer/LinkRenderer';
-export { Logo } from './Logo';
+export { Logo } from './Logo/Logo';
 export { ProfileImg } from './ProfileImg';
-export { Textarea } from './Textarea';
-export { TextInput } from './TextInput';
+export { Textarea } from './Textarea/Textarea';
+export { TextInput } from './TextInput/TextInput';
 export { Toast, useToast } from './Toast';
 export { Checkbox, type CheckboxProps } from './Checkbox';
-export { GoToTopButton } from './GoToTopButton';
-export { CreateButton } from './CreateButton';
-export { Carousel } from './Carousel';
-export { StatusTag } from './StatusTag';
+export { GoToTopButton } from './GoToTopButton/GoToTopButton';
+export { CreateButton } from './CreateButton/CreateButton';
+export { Carousel } from './Carousel/Carousel';
+export { StatusTag } from './StatusTag/StatusTag';
 export {
   UserActionMenu,
   type UserActionMenuProps,
 } from './UserActionMenu/UserActionMenu';
-export { Footer } from './Footer';
+export { Footer } from './Footer/Footer';
 export { KeywordChip } from './KeywordChip/KeywordChip';
-export { ReviewItemBar } from './ReviewItemBar';
+export { ReviewItemBar } from './ReviewItemBar/ReviewItemBar';
 export {
   Accordion,
   type AccordionItem,
   type AccordionProps,
 } from './Accordion/Accordion';
+export { ToastProvider } from './Toast/ToastContext';
+export { SpeechBubble } from './SpeechBubble/SpeechBubble';
