+++ src/components/Molecules/EmptyState/MyPageEmptyState.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from '@/components';
+import { Button } from '@/components/Atoms';
 import {
   type MainTabType,
   type SubTabType,
