+++ src/components/Molecules/ErrorPage/Errorpage.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components';
+import { Button } from '@/components/Atoms';
 import Image from 'next/image';
 
 export const ErrorPage = () => {
