+++ src/components/Molecules/ErrorPage/ServerErrorPage.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components';
+import { Button } from '@/components/Atoms';
 import Image from 'next/image';
 
 export const ServerErrorPage = () => {
