+++ src/components/Molecules/index.ts
@@ -9,15 +9,18 @@ export {
   Line,
   MainCard,
   type CardProps,
-} from './Card';
-export { Dropdown, type DropdownProps } from './Dropdown';
-export { ErrorPage, ServerErrorPage } from './ErrorPage';
-export { Header } from './Header';
+} from './Card/Card';
+export { Dropdown, type DropdownProps } from './Dropdown/Dropdown';
+export { ErrorPage } from './ErrorPage/Errorpage';
+export { Header } from './Header/Header';
 export {
   MainEmptyState,
-  MyPageEmptyState,
   type MainEmptyStateProps,
-  type MyPageEmptyStateProps,
-} from './EmptyState';
-export { SideButtonSection } from './SideButtonSection';
-export { ProfilePopover } from './ProfilePopover';
+} from './EmptyState/MainEmptyState';
+export { SideButtonSection } from './SideButtonSection/SideButtonSection';
+export { ProfilePopover } from './ProfilePopover/ProfilePopover';
+export { Modal } from './modal/modal';
+export { useModal } from './modal/hooks/useModal';
+export { SearchInput } from './Search/SearchInput';
+export { MyPageEmptyState } from './EmptyState/MyPageEmptyState';
+export { ServerErrorPage } from './ErrorPage/ServerErrorPage';
